api-gateway/requirements.txt
fastapi[standard]==0.118.0
grpcio==1.75.1
grpcio-tools==1.71.0
uvicorn==0.37.0
protoletariat==3.3.10
requests==2.32.5